[REFACTORING] Переместил мапперы к своим классам
