--- app/src/main/java/com/desuzed/everyweather/data/network/dto/ApiTypeError.kt
@@ -1,6 +1,0 @@
-package com.desuzed.everyweather.data.network.dto
-
-sealed class ApiTypeError{
-    open class WeatherApi : ApiTypeError()
-    open class OpenWeatherApi : ApiTypeError()
-}
