--- app/src/main/java/com/desuzed/everyweather/data/network/dto/ApiTypeWeather.kt
@@ -1,6 +1,0 @@
-package com.desuzed.everyweather.data.network.dto
-
-sealed class ApiTypeWeather{
-    open class WeatherApi : ApiTypeWeather()
-    open class OpenWeatherApi : ApiTypeWeather()
-}
