+++ app/src/main/java/com/desuzed/everyweather/data/network/dto/weatherApi/DayDto.kt
@@ -1,5 +1,7 @@
 package com.desuzed.everyweather.data.network.dto.weatherApi
 
+import com.desuzed.everyweather.model.model.Day
+import com.desuzed.everyweather.util.EntityMapper
 import com.google.gson.annotations.SerializedName
 
 class DayDto {
@@ -38,4 +40,20 @@ class DayDto {
     var uv = 0f
 
     var pop : Int = (popRain + popSnow)/2
+}
+
+class DayMapper : EntityMapper<DayDto, Day> {
+    override fun mapFromEntity(entity: DayDto): Day {
+        return Day(
+            entity.maxTemp,
+            entity.minTemp,
+            entity.maxWind,
+            entity.totalPrecip,
+            entity.avgHumidity,
+            entity.popRain,
+            entity.popSnow,
+            entity.conditionDto?.text.toString(),
+            entity.conditionDto?.icon.toString()
+        )
+    }
 }
