+++ app/src/main/java/com/desuzed/everyweather/data/network/dto/weatherApi/ForecastDayDto.kt
@@ -1,5 +1,8 @@
 package com.desuzed.everyweather.data.network.dto.weatherApi
 
+import com.desuzed.everyweather.model.model.ForecastDay
+import com.desuzed.everyweather.model.model.Hour
+import com.desuzed.everyweather.util.EntityMapper
 import com.google.gson.annotations.SerializedName
 
 class ForecastDayDto {
@@ -17,4 +20,20 @@ class ForecastDayDto {
 
     @SerializedName("hour")
     var hourDto: ArrayList<HourDto>? = null
+}
+
+class ForecastDayMapper : EntityMapper<ForecastDayDto, ForecastDay> {
+    override fun mapFromEntity(entity: ForecastDayDto): ForecastDay {
+        val listHour = mutableListOf<Hour>()
+        entity.hourDto?.forEach {
+            listHour.add(HourMapper().mapFromEntity(it))
+        }
+        return ForecastDay(
+            entity.date,
+            entity.dateEpoch,
+            DayMapper().mapFromEntity(entity.day!!),
+            AstroMapper().mapFromEntity(entity.astroDto!!),
+            listHour
+        )
+    }
 }
