+++ app/src/main/java/com/desuzed/everyweather/data/network/dto/weatherApi/HourDto.kt
@@ -1,5 +1,7 @@
 package com.desuzed.everyweather.data.network.dto.weatherApi
 
+import com.desuzed.everyweather.model.model.Hour
+import com.desuzed.everyweather.util.EntityMapper
 import com.google.gson.annotations.SerializedName
 
 class HourDto {
@@ -54,4 +56,18 @@ class HourDto {
 
     @SerializedName("gust_kph")
     var gust = 0f
+}
+
+class HourMapper : EntityMapper<HourDto, Hour> {
+    override fun mapFromEntity(entity: HourDto): Hour {
+        return Hour(
+            entity.timeEpoch,
+            entity.temp,
+            entity.conditionDto?.text.toString(),
+            entity.conditionDto?.icon.toString(),
+            entity.windSpeed,
+            entity.windDegree,
+            entity.pressureMb
+        )
+    }
 }
