+++ app/src/main/java/com/desuzed/everyweather/data/network/dto/weatherApi/LocationDto.kt
@@ -1,5 +1,7 @@
 package com.desuzed.everyweather.data.network.dto.weatherApi
 
+import com.desuzed.everyweather.model.model.Location
+import com.desuzed.everyweather.util.EntityMapper
 import com.google.gson.annotations.SerializedName
 
 class LocationDto {
@@ -19,12 +21,27 @@ class LocationDto {
     var lon = 0f
 
     @SerializedName("tz_id")
-    var tzId: String= ""
+    var tzId: String = ""
 
     @SerializedName("localtime_epoch")
     var localtime_epoch: Long = 0
 
     @SerializedName("localtime")
-    var localtime: String= ""
+    var localtime: String = ""
 
+}
+
+class LocationMapper : EntityMapper<LocationDto, Location> {
+    override fun mapFromEntity(entity: LocationDto): Location {
+        return Location(
+            entity.name,
+            entity.region,
+            entity.country,
+            entity.lat,
+            entity.lon,
+            entity.tzId,
+            entity.localtime_epoch,
+            entity.localtime
+        )
+    }
 }
