+++ app/src/main/java/com/desuzed/everyweather/data/network/dto/weatherApi/WeatherResponseDto.kt
@@ -1,9 +1,11 @@
 package com.desuzed.everyweather.data.network.dto.weatherApi
 
-import com.desuzed.everyweather.data.network.dto.ApiTypeWeather
+import com.desuzed.everyweather.model.model.ForecastDay
+import com.desuzed.everyweather.model.model.WeatherResponse
+import com.desuzed.everyweather.util.EntityMapper
 import com.google.gson.annotations.SerializedName
 
-class WeatherResponseDto : ApiTypeWeather.WeatherApi()  {
+class WeatherResponseDto {
     @SerializedName("location")
     var locationDto: LocationDto? = null
 
@@ -13,5 +15,17 @@ class WeatherResponseDto : ApiTypeWeather.WeatherApi()  {
     @SerializedName("forecast")
     var forecastDto: ForecastDto? = null
 
+}
+
+class WeatherResponseMapper : EntityMapper<WeatherResponseDto, WeatherResponse> {
+    override fun mapFromEntity(entity: WeatherResponseDto): WeatherResponse {
+        val location = LocationMapper().mapFromEntity(entity.locationDto!!)
+        val current = CurrentMapper().mapFromEntity(entity.currentDto!!)
+        val forecastDayList = mutableListOf<ForecastDay>()
+        entity.forecastDto?.forecastday?.forEach {
+            forecastDayList.add(ForecastDayMapper().mapFromEntity(it))
+        }
+        return WeatherResponse(location, current, forecastDayList)
+    }
 }
