--- app/src/main/java/com/desuzed/everyweather/data/network/dto/weatherApi/mappers/WeatherResponseMapper.kt
@@ -1,17 +1,0 @@
-package com.desuzed.everyweather.data.network.dto.weatherApi.mappers
-
-import com.desuzed.everyweather.model.model.ForecastDay
-import com.desuzed.everyweather.model.model.WeatherResponse
-import com.desuzed.everyweather.data.network.dto.weatherApi.WeatherResponseDto
-
-class WeatherResponseMapper: EntityMapper<WeatherResponseDto, WeatherResponse> {
-    override fun mapFromEntity(entity: WeatherResponseDto): WeatherResponse {
-        val location = LocationMapper().mapFromEntity(entity.locationDto!!)
-        val current = CurrentMapper().mapFromEntity(entity.currentDto!!)
-        val forecastDayList = mutableListOf<ForecastDay>()
-        entity.forecastDto?.forecastday?.forEach {
-            forecastDayList.add(ForecastDayMapper().mapFromEntity(it))
-        }
-        return WeatherResponse(location, current, forecastDayList)
-    }
-}
