+++ app/src/main/java/com/desuzed/everyweather/data/repository/RemoteDataSource.kt
@@ -1,6 +1,5 @@
 package com.desuzed.everyweather.data.repository
 
-import com.desuzed.everyweather.data.network.dto.ApiTypeWeather
 import com.desuzed.everyweather.data.network.dto.weatherApi.ErrorDtoWeatherApi
 import com.desuzed.everyweather.data.network.dto.weatherApi.WeatherResponseDto
 import com.desuzed.everyweather.data.network.retrofit.NetworkResponse
@@ -13,7 +12,6 @@ class RemoteDataSourceImpl : RemoteDataSource {
      * Default language: English
      */
     var lang: String = "en"
-    var mApiType: ApiTypeWeather = ApiTypeWeather.WeatherApi()
 
     override suspend fun getForecast(query: String): NetworkResponse<WeatherResponseDto, ErrorDtoWeatherApi> =
         withContext(Dispatchers.IO) {
@@ -21,22 +19,6 @@ class RemoteDataSourceImpl : RemoteDataSource {
                 .getInstance()
                 .getForecast(query, lang)
         }
-
-
-//    suspend fun getForecast1(query: String): NetworkResponse<ApiTypeWeather, ApiTypeError> {
-//        //        if (mApiType is ApiType.WeatherApi) {
-////            return WeatherApiService
-////                .getInstance()
-////                .getForecast(query, lang)
-////        } else if (mApiType is ApiType.OpenWeatherApi) ..... запускаем другой сервис
-//        return WeatherApiService
-//            .getInstance()
-//            .getForecast(query, lang)
-//    }
-
-    /**
-     * Here will be calls from other weather APIs
-     */
 }
