+++ app/src/main/java/com/desuzed/everyweather/util/EntityMapper.kt
@@ -1,4 +1,4 @@
-package com.desuzed.everyweather.data.network.dto.weatherApi.mappers
+package com.desuzed.everyweather.util
 
 interface EntityMapper <Entity, DomainModel> {
     fun mapFromEntity (entity: Entity) : DomainModel
