+++ app/src/main/java/com/desuzed/everyweather/view/fragments/LocationFragment.kt
@@ -18,14 +18,14 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.RecyclerView
 import com.desuzed.everyweather.App
 import com.desuzed.everyweather.R
-import com.desuzed.everyweather.view.adapters.FavoriteLocationAdapter
-import com.desuzed.everyweather.databinding.FragmentLocationBinding
+import com.desuzed.everyweather.data.network.ActionResultProvider
 import com.desuzed.everyweather.data.room.FavoriteLocationDto
+import com.desuzed.everyweather.data.room.FavoriteLocationMapper
+import com.desuzed.everyweather.databinding.FragmentLocationBinding
 import com.desuzed.everyweather.model.vm.AppViewModelFactory
 import com.desuzed.everyweather.model.vm.SharedViewModel
-import com.desuzed.everyweather.data.network.ActionResultProvider
 import com.desuzed.everyweather.view.MainActivity
-import com.desuzed.everyweather.data.network.dto.weatherApi.mappers.FavoriteLocationMapper
+import com.desuzed.everyweather.view.adapters.FavoriteLocationAdapter
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import java.util.*
